Had 2 index.js files. Rename one to apiRoutes.js to avoid confusion. Let VS Code automatically update import statements.
