+++ src/routes/index.js
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 const router = Router();
-import apiRoutes from './api/index.js';
+import apiRoutes from './api/apiRoutes.js';
 import htmlRoutes from './htmlRoutes.js';
 router.use('/api', apiRoutes);
 router.use('/', htmlRoutes);
